[gapless] Fix rating resets (bgo#614863)
When UpdateLastPlayed is called from OnAboutToFinish, it saves the
currently played track to the database. The problem is that the playing
track and the rated track are different objects, the playing track
still has the original rating value.

This commit refreshes the track before updating and saving, which
brings its properties up-to-date.