[XOverlay] fix video embed bnc#585817, bgo#548635
GStreamer requires an X11 window handle to be set on the XOverlay
interface of the video sink in order to render video. If one is not set
in time, GStreamer will create its own X window. Closing this window
will result in a crash, since it becomes unrealized, after which
GStreamer will continue to try to render to it.

Before, the X11 video window was created when the Now Playing UI
was realized. The window is now created when the pipeline is created, by
raising the new PrepareVideoWindow event. Technically, this event should
be raised when the prepare-xwindow-id message is pushed onto the
pipeline bus, but this is raised out of the main thread.