[TrackInfoDisplay] support forcing size/spacing
Allow overriding the artwork size with an explicit size, instead of the
default size which is determined by the allocation height. This ensures
that regardless of the allocation, we pull a safe artwork size (e.g.
from valid scale cache).

Additionally allow for configuring the space between text and artwork.