[Dap] Fix not transcoding some files when we should
When a source file to be transferred to a DAP did not match any
available profile, it was not transcoded due to the fact that the
extension returned was null and many in AcceptableMimeTypes for a
device suffered the same. In that case null == null and no transcoding
was performed.

https://bugzilla.gnome.org/show_bug.cgi?id=633764

Signed-off-by: Gabriel Burt <gabriel.burt@gmail.com>