AppleDevice: Remove invalid tracks before syncing (bgo#634652)
It can happen that the device has tracks that are in the database
but not as physical files. In this case the files have a null IpodPath
property, which was causing an NRE at the moment of loading the device
in Banshee. These tracks couldn't be seen anyway in the device's
interface, so it is safe to remove them (besides, it seems the that
the existence of them would actually prevent the sync process
from really happen properly).