X11NotificationArea: Enhance check for RGBA visual (bgo#638477)
Compare the sum of the precisions of all colors with the depth of the
visual.

This uses functions that are only available in gdk-x11 >= 2.22,
so we only do it in this case. Users with an older version of gdk-x11
will still have the bug, but it improves the situation quite a bit.

Signed-off-by: Bertrand Lorentz <bertrand.lorentz@gmail.com>