PlayQueueSource: Fix high memory usage when shuffling (bgo#676144)
We were using Enumerable.Range by passing as a second parameter the
value of the last integer in the range we want. But it is in fact the
number of integers in the range.

So for example shuffling tracks with a ViewOrder between 1000 and 1003
would allocate a list of 1003 integers, instead of just the 3 we need.