DbIteratorJob: Abort job after several consecutive failures (bgo#673416)
An exception while executing a job (in IterateCore) probably means the
database entry will still be returned by the next execution of the
query. This causes us to go into a tight loop of "Get entry, try to
process it, fail".

We now stop the job after 5 consecutive exceptions raised by
IterateCore. That could mean all items have not been processed by the
job, but it's much better than eating all the CPU...