LastFm: Null check for scrobble parameters (bgo#685920)
As noted in bgo#685920 an exception was being thrown attempting to escape
a null scrobble parameter when building Post data. This was preventing
scrobbles from being submitted. This is a regression in version 2.6.0 due
to the new way we build the submissions for the Last.FM 2.0 API.

We will now perform a null check on each parameter and only attempt to
escape it if the value is not null.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>