build: make GIO backend be enabled by default (bgo#686584)
Other IO backends were deprecated some time ago, so now the only option
is Gio. Before this happened, the backend to build was guessed at config
time by our M4 macros.

Problem is, if dependencies for GIO were not met, the default was to
silently not build any IO backend, which makes Banshee lacks lots of
useful features.

This commit changes this so it is now mandatory to pass the --disable-gio
flag if the user wishes to compile without any IO backend.

It doesn't completely solve bgo#686584 but helps alleviate the problem.
And has the added benefit of making more versions of dependencies be
clearly visible at the top of the macro.