PrimarySource: prevent AddTrackJob respawn if cancelled (bgo#684089)
When the add_track_job object got cancelled by the user, the method
IncrementAddedTracks was resetting it to null after having checked that
its IsFinished property was True. IsFinished returns True if the job has
finished normally and if it was cancelled. For the former, this code flow
would work because it would be set to null in the last iteration, and
would not be queried again. But in the case of a cancellation, next
iteration of the foreach would check if (AddTrackJob.IsCancelRequested),
which would create a new UserJob again (as AddTrackJob is a property that
wraps add_track_job but creates one if it is null).

The best and most simple way to fix this is simply never reset
add_track_job to null until the loop has finished.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>