PrimarySource: fix crash if 1st track had a Dap transcoding error (bgo#700753)
The method IncrementAddedTracks() was accessing the field add_track_job, which
would make Banshee crash with a NullReferenceException. By judging from the
stack trace that the user supplied, this was likely to happen only if the very
1st track being transferred through Dap had a transcoding error (otherwise the
code flow would access a method which initializes add_track_job if it is null).

By using the property AddTrackJob instead of the field add_track_job, we make
this method capable of initializing the field too in case it's reached the first
when Dap transfer happens.