build: Rework build configuration for GStreamer backends
Using a PLAYBACK_BACKEND variable in SUBDIRS is incorrect, as it
always causes one of the backends to be omitted from the source tarball,
thus breaking "make distcheck".

The two GStreamer backends can now be built conditionnaly, with the
appropriate --enable-gst-sharp and/or --disable-gst-native configure
option.

By default, only the native backend is built, as was the case before.
You can choose to build both backends, or none if you want Banshee to be
very silent...

An indication whether the native backend will be built is added to the
configure output, for clarity.