build: Use notify-sharp-3.0 when present on the system
Now that notify-sharp has a 3.0 release that depends on GTK 3, we can
use it and revert back to the same behavior we had in the GTK 2 world:
use the appropriate system-installed notify-sharp if present, and if
not, use our own internal copy.

This removes the --with-system-notify-sharp configure option, which
nobody should have been using anyway.