build: default to build just one media backend
Unless explicitly required by the user at configure time, the build
should default to only enable one media backend, not two at the same
time.

Banshee can run with more than one media backend, but this feature is
mainly intended for the use case of having two backends with different
capabilities. However, in this case we're talking about selecting the
managed or the unmanaged backend, and both (should) have the same
capabilities.

Before this patch, the user had to explicitly enable the managed
backend and disable the native one if she wanted to run the managed
one without loading the native one. This is now corrected: if the
user doesn't supply any option, only the native backend is built,
and if the user supplies the --enable-gst-sharp option, only the
managed one is built.