WebBrowser: Depend on libsoup >= 2.42, drop libsoup-gnome dependency
Since 2.42, libsoup-gnome is deprecated and all features are now
provided by libsoup itself, with saner defaults. So we now require
version 2.42, which allows us to drop the dependency on libsoup-gnome.

Note that we still use the GNOME proxy settings when available, as this
is now done by default by libsoup itself.