Gio: prevent NRE when scanning for CD device (bgo#721203)
This is a similar fix to the one for BGO#699735 which consists
in detecting a null value earlier before creating an instance
of UdevMetadataSource class.

This time we also throw ArgumentNullException now in the ctor
of this class to fail faster when this situation happens,
instead of failing when some code queries any property of
an object of this class which happened to wrap a null value.

[1] https://git.gnome.org/browse/banshee/commit/?h=stable-2.6&id=1f8d8746eb122449a3489901a87f8a15e54ae4cf