Gio: bail from foreach iteration after finding null (bgo#721203)
Previous commit [1] didn't take in account that a 'yield return'
is not equivalent to a 'return' statement with regards to loops
(the latter will exit the loop, but not the former). So we need
to pass to the next iteration of the loop manually via `continue`
in every case where we want to discard creating a new Volume
object because of finding unwanted null collaborators.

https://git.gnome.org/browse/banshee/commit/?id=20bacf9a3d06b2d136b474319f0b8224947bc4b8