GStreamerSharp: Rework initialization to match the native backend
The native C GStreamer backend builds the pipeline during the "delayed
initialization" phase, which allows other extensions to be notified when
it is done. So we move the corresponding part from the Gst# PlayerEngine
constructor to the Initialize method.

This fixes the "Equalizer" menu item to be enabled when an equalizer is
available, and maybe some other things.