GStreamerSharp: Fix equalizer crashes
For the equalizer we're using the "equalizer-10bands" element, which
always has 10 bands and hard coded values for the band frequencies and
gain range. So we use these values directly, instead of going through
child elements, which causes crashes, and might not be possible anymore.

We could use the "equalizer-nbands" to be generic, but I don't see the
point right now.