build: require NUnit 2.5
Previous commit [1] stated something which is not always true
actually: although mono installs a mono-nunit.pc that contains
the mono version (as opposed to NUnit version), some distros
(like Debian/Ubuntu) have made that pc file be a symlink from
/usr/lib/pkg-config/mono-nunit.pc to /usr/lib/pkg-config/nunit-pc.

So the best thing is avoid this mess and only depend on NUnit,
which BTW brings some nice API since version 2.5 (mono embeds
version 2.4.8 [2]):

  Assert.That(actual, Is.EqualTo(expected));

This API is more readable (maps better to English language) and has
the added benefit of providing a way to never confuse again the
parameter *actual* with *expected* that was so easy to switch by
mistake on the older Assert.AreEqual(x,y) syntax.

Some lines are added to the HACKING file about this as well, and
the last test I wrote is converted to this syntax, as the first
real sample to have in the codebase.

[1] https://git.gnome.org/browse/banshee/commit/?id=3891ed1e5aaa7d06497525ab5bfc44110939a369
[2] https://github.com/mono/mono/blob/master/mcs/nunit24/CommonAssemblyInfo.cs