tests: bail early if nunit-console is not found (bgo#722565)
It's not worth checking for this at configure time because
one can build Banshee fine when having nunit libraries but
not the runner (i.e.: in a Debian-based distro, having the
package "libnunit-cil-dev" installed, but not the one
named "nunit-console").

But we need to check before running the tests, otherwise
the user may get many errors (as many as the number of
test assemblies we have), instead of just one, making it
very clear what is missing.