MediaEngineTests: bring assertion exceptions to main thread
The AssertTransition() method in the tests was connecting a
PlayerEventHandler to the PlayerEngineService being tested
but it didn't have any try-catch around it, which means that
any generated AssertionException caused by a test not passing
would really be ignored (just printing a warning on the
terminal instead of making the NUnit test fail).

A new guarded_handler is now created, which calls the old one,
and which has a try-catch, which would set any exception caught
to a variable of the method, which would then later be captured.

The change of the type of the 'handler' variable is not really
a required change here, but doing it allows us to have the
compiler complain to us if we try to disconnect the old handler
instead of the new one called 'guarded_hanlder'.