MediaEngineTests: honor String.Format() usage in assertion msg
This assertion message was using String.Format() when in fact
there was no need to do so, as there was no parameter interpolation
at all. However, instead of just removing the call to String.Format
we can improve this slightly to include the number of seconds
waited as a parameter. This way, in case this number changes in
the future, we will only need to change one line and not two.

The last reason to use 3 seconds is mentioned in this commit:
https://git.gnome.org/browse/banshee/commit/?id=2212116a814640ffce06ae2fe0fb8f00a0de10f0