MediaEngineTests: improve information about failed tests
The AssertionException recently captured in previous commit had
line numbers referring to the exact place in the
Action<PlayerEventArgs> handler where it failed, but its stack
trace didn't go deep enough to tell us which lines the callers
of AssertTransition come from.

The way to fix it is to create a new exception object in the
non-handler thread, which copies the exception message of the
captured assertion exception (and stores it as inner exception
too, just in case).

We also add a message to the Asserts in the handler in order
to distinguish better if we're comparing an event or a state.