GStreamerSharp: PlayerEngine: avoid spurious *Stream events
The managed backend wasn't checking the player state before sending
{End|Start}OfStream events at GStreamer's StreamStart signal, like
the native backend was doing. This was making the MediaEngineTests
fail for this backend.

With this, the tests now pass (there's only a TearDown failure left
which is unrelated to this).

(Note: the player states to check recently had to change in the
native backend [1], so we accommodate the managed backend here
too, accordingly.)

https://git.gnome.org/browse/banshee/commit/?id=6c890659b766879e0bbd08e6ea29d836b9c3281c