MediaEngineTests: make test for bgo#679938 compatible with stable branch
The stable branch can be run against GStreamer 0.10, which for
some reason sends NextTrackStarting events when the current
track finishes, even if there's no next track. As this test
doesn't care about this signal but about the AboutToFinish one,
we tweak the test to only care about the events that matter
for the test.

Also, somehow this lambda syntax didn't seem to compile while
in the master branch it worked:

  MCS   ../../../bin/Banshee.Services.dll
./Banshee.MediaEngine/Tests.cs(154,1): error CS1660: Cannot convert `lambda expression' to non-delegate type `object'
./Banshee.MediaEngine/Tests.cs(154,1): error CS1660: Cannot convert `lambda expression' to non-delegate type `object'

So a System.Action variable is now used instead...

This commit will be forward-ported to the master branch from the
stable branch, to keep the code in sync.