DBus: don't try to connect in GrabDefaultName() if DBus disabled
If DBus was not enabled, we were properly returning true (as in
primary_instance == true) but we were trying to connect when
we should not.

Also, by changing primary_instance to be initialized to true,
instead of false, at first, then we need to change its value in
less places.