configure: check GTK# before GStreamer#
Some people have already asked how to build with the native backend
instead of the managed one, just with the hope of avoiding the new
bleeding-edge dependency (because by trying to compile GStreamer#
they are asked to install the GLib# dependency, and they think that
they will find yet another level of yak shaving if they go on...).

By bringing up GTK# macro before the GStreamer# one (which checks
for GLib# too), we will reduce one level of depth in dependency
searching strategies of our potential contributors (because GLib#
will be already installed when they try to install GStreamer#).