GStreamerSharp: bring back option to disable gapless
Our current unmanaged backend still has this option, so we should
match the new managed backend to the same expectations. This brings
back some code paths (and comments) from the unmanaged backend that
were not present in the managed one (namely the use of the pending_*
and next_track_pending fields).

We plan to remove the Gapless setting soon (so it is always enabled
by default and can't be changed), but we're not quite ready yet,
given the number of bugs that are still affected by this setting (we
have fixed some recently, which were backported to 2.6.2, but we're
not there yet). We will track this progress in bgo#638943.

Even though disabling gapless is an option that is not perfect
(right now I'm experiencing an issue with volume being changed at
transition), it is good to have it until our first 3.0 RC, so
we can ask users to tell us if any given problems they
experience with the new managed backend are gapless-related or
not. We should concentrate in fixing gapless bugs rather than
non-gapless issues that this brings.