DbImportManager: fix out-of-sync collection while iterated (bgo#724635)
As reported in the stacktrace of bug 724635, it could happen that
while syncing an iPod, the NotifyAllSources() method could crash
at some iteration of its foreach loop because the collection being
iterated had been modified since the loop started.

By making a copy of the collection before iterating, we avoid this
potential problem.

(Backported manually from master branch, as it could not be simply
cherry-picked, because it would break the build.)