Gio: don't try to mount null devices (bgo#726147)
The log in bgo#726147 can tell us that we were forgetting a null
check in HandleMonitorMountAdded(). Turns out we had a similar
null check a bit later in this file in the VolumeRemoved() method,
which I copy-pasted here just by replacing the "unmount" word in
the log statement with "mount" in this case.