HACKING: add missing rule about parameter names
There's a mixture of style in Banshee's codebase about param
names. Sometimes they are written in the same way vars/fields
are written (under_score notation), but sometimes they use
camelCaseNotation. Lately I've come to realize that the latter
is a bit more frequent, so we should write this down ASAP
before we grow the mixed-style code indefinitely.