Mtp: Ignore mount/unmount events to fix device load (bgo#729438)
When plugging an MTP device into the system, banshee must
unmount it to be able to use it. Since the hardware manager
emits signals when a device is unmounted, this causes banshee
to unmap a device that it is in the middle of configuring.

This patch changes the signals that GIO Manager watches on
GNU/Linux systems from VolumeAdded, MountAdded and MountRemoved
to VolumeAdded and VolumeRemoved. This has the added benefit
that users with automount disabled have their devices recognised.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>