build: simplify uninstall-hook target in root folder
The way to check if $(pkglibdir) exists can be simplified to just
check for its existence now. I guess `find` was used in combination
with `wc` before because it could happen that there were leftover
subfolders not properly uninstalled, which would make `rmdir` error
out.

However, this situation has been fixed in [1], so we can now
assume that $(pkglibdir) doesn't contain any elements when this
hook happens (if it contains any elements, it will be a bug in the
build infrastructure, so with this `rmdir` would fail fast, which
is a good thing).

[1] https://git.gnome.org/browse/banshee/commit/?id=50257fafdc0077754cb2cb155db048261fdfbb73