Refactoring: rename "SelectedTracks" methods&labels (bgo#726338)
Renamed all "...Tracks" to "SelectedTracks" labels and methods,
following the previous commit, which places all the actions under
this "Selected Track(s)" submenu.

This is especially useful, to distinguish them from the next
methods and actions which will be committed in the next changeset,
which will introduce a new Playing Track submenu.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>