ThickClient: Add "Playing Track" submenu (bgo#726338)
When some user requested to have CoverArt modification for the
playing track in the "Edit" menu (bgo#724795), we realised that
it was not clear at all that this menu affected just to selected
tracks, not to the playing track. The two previous patches then
modified the UI to include these options under a new "Selected
Track(s)" submenu.

Now, this change introduces a similar submenu (with the same
options) but which applies only to the playing track. This only
works for now if the track played is in the main Music library,
and it can be extended for more sources in the future.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>