Gio: protect call to event delegates in HardwareManager
Make Gio's HardwareManager class respect this concurrency rule
that is already a good practice held in other parts of the
codebase. (For a good explanation, you can check [1])

This commit also makes the event handling more consistent to
how it is done in other parts of this class (having two overloads
of the same method, with different parameter types).

[1] http://www.mono-project.com/Gendarme.Rules.Concurrency#ProtectCallToEventDelegatesRule

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>