Dap.MassStorage: add logging to empty catch{} blocks
While finding a fix for bug 729438 we have gone through some
scenarios in which an exception was being thrown, and caught
by empty catch{} blocks, hiding a regression which would have
been a bit clear if there was at least some logging in these
code flows.

So let's add it. If it becomes too noisy, we could still move
it to only be logged on debug mode, but never logging it looks
like a mistake.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>