SourceRowRenderer: Rework rendering of the cell background
If the cell is selected, just let the GTK theme do the drawing.

If the cell is highlighted but not selected, like when you right click
on a non-selected cell that is not selected, draw a simple 1 pixel
border, like we did before.

The notification animation is now a simple rectangle fill, using a color
halfway between normal and selected. We can't use the color of the
"active" state anymore, because it now seems to be the same as the
normal color, at least for treeviews.

We should probably use GTK+ animations for that, but I still need to
figure out how...

Also update Hyena to the latest git master, to bring in all the latest
ListView changes.