ThickClient: workaround for invisible (instead of bold) text (bgo#732838)
After several hours attempting to find a reduced testcase for this
problem (which appears to be somewhere deeper in the stack: pango,
cairo, or gtk), I've found this harmless workaround that actually
makes the first line of text (in TrackInfoDisplay class) be more
consistent with the second line of text (the latter already wrapped
other markup elements with a wider <span> element using the default
text colour).

The TextColor and BackgroundColor properties seem to have wrong
values though (both black, rgb#000000), compared to the stable
branch (GTK2), but that will be tracked in a different bug, because
it is unrelated to this issue.