TrackInfoDisplay: proper mid color for 'from' and 'by' (bgo#732898)
The mid-color between the default text color and the background color
of the widget was not being calculated properly because the background
color obtained for the TrackInfoDisplay widget was the same as the text
color. This behaviour must have changed recently in GTK3, but what we
are sure about is that using the Parent's reference colors we get the
proper colors again (probably because the parent of this widget would
be a ToolItem, inside a Toolbar, which has its background color surely
defined, compared to TrackInfoDisplay widget which is implemented in
the Banshee codebase.