TrackInfoDisplay: proper fix for not bold but invisible text (bgo#732838)
OnStyleUpdated() was in charge of filling the color fields of
TrackInfoDisplay class with values, but it could happen that
these colors were retrieved before OnStyleUpdated() hadn't been
run yet even once, which resulted in a black color, but all
transparent (alpha=0). The way to fix it is make the colors be
properties that query the proper value on demand if the value
had not been set yet, or is resetted.

Thanks to Bertrand Lorentz for the heads up about the hint of
what was really going on badly.