Logging: bring latest hyena (& kill majority of warnings from it)
Latest change in hyena is the deprecation of Log.Exception() methods
because of its severity ambiguity. To fix the warnings generated by
this deprecation, I've taken this rule of thumb:

- If the exception being logged comes from a generic catch(Exception)
it's probably a way to avoid that Banshee crashes, and having a generic
Exception type (instead of a more specific one, i.e. FormatException)
most likely means that the reason for the exception is unknown, so it's
convenient to file this as an error. If any of this particular cases
becomes frequent enough, we can always look at the stacktrace, and put
a more specific catch() block that uses Log.Warning() in the future.

- If the exception being logged comes from a specific catch block of
a specific exception type (derived from System.Exception) it normally
corresponds to a non-severe problem that can be ignored, but logged as
a warning.