CoverArt: display embedded artwork even if tags missing (bgo#692107)
Most MetadataProvider classes cannot retrieve cover art if there
is no album name information in the track because they get retrieve
the metadata from online sources. However, there is a provider which
does it locally: EmbeddedMetadataProvider.

In this case, the restriction about album name being known doesn't
really apply, so let's move this restriction to the online providers
and remove it from the more generic MetadataService and CoverArtJob
classes.